refactor: change example imports
